Add daemonx (extended daemon) example.
This is an attempt to rework the daemon example so that it becomes more useful. It also serves as a showcase utilizing most ctlra features, and as a hacking playground to explore the capabilities of the current flock of NI Maschine devices.

In particular, this version of the daemon adds the capability to output text on the screen, improves the local feedback, and also adds optional MIDI and MCP feedback, including support for scribble strips, timecode and meter displays. For good measure, it also throws in some external configuration capabilities and a screenshot utility via custom sysex messages. It also adds a Ctlra prefix to the ALSA client name, so that it can be distinguished more easily from the dummy MIDI device created by ALSA itself.

NOTE: This hasn't been tested with anything but the Maschine Mk3 yet; it hopefully works at least to some extent with the other devices supported by ctlra, but probably needs some (or a lot of) work, depending on the particular device. Also, there's no mappa support yet, so to actually make it work as a kind of Mackie emulation with the Maschine Mk3, you currently need to use midizap (https://github.com/agraef/midizap) with the Maschine.midizaprc configuration.